Move wrappers to common wrappers
